+++ server/api-server/src/main.ts
@@ -30,6 +30,7 @@ async function bootstrap() {
     methods: 'GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS',
     credentials: true,
   });
+
   const session = getSession();
   app.use(session);
   app.use(passport.initialize());
