+++ server/api-server/src/streams/streams.controller.ts
@@ -5,29 +5,13 @@ import {
   Patch,
   Param,
   Query,
-  UseGuards,
-  Session,
-  HttpException,
-  HttpStatus,
 } from '@nestjs/common';
 import { StreamsService } from './streams.service';
 import { UpdateStreamDto } from './dto/update-stream.dto';
-import { LoggedInGuard } from '../auth/guard/logged-in.guard';
 
 @Controller('streams')
 export class StreamsController {
   constructor(private readonly streamsService: StreamsService) {}
-
-  @UseGuards(LoggedInGuard)
-  @Get('me')
-  async getStreamKey(@Session() session: Record<string, any>) {
-    if (!session.userId) {
-      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
-    }
-
-    const streamKey = await this.streamsService.getStreamKey(session.userId);
-    return { streamKey };
-  }
 
   @Get()
   findAll(@Query('page') page = '1', @Query('size') size = '5') {
